Essentially fixed camera jittering
Fixed camera jittering as much as needed for our purposes. I removed the factor of 10 scaling the animation, which fixed the problems. It also didn't seem to be necessary, really. In theory there's probably still some issue with how the velocity is always snapped to a constant length, but this seems fine for now.