Fixed behaviours
Discovered that the behaviours were wrong (flock centering and velocity matching), so I fixed them (they were generating arbitrary velocities and not really accelerations. It's not clear that it's better, but it is technically correct. Reduced the acceleration calculations to every 20 frames to avoid jittering, which I think is a better approach than relying on fine tuning. Ideally we would smooth the accelerations somehow, or have them apply over a longer time.
Works and does create a nice animation (except that the boids pass through buildings)