Building collisions partially working
Collisions now use handwritten cylinder ray casting (I had tried the Blender ray_cast API up until a few weeks ago, and it was impossible to use). It works a little – you can see the boids trying to fly around the buildings. But unfortunately I couldn't get it completely functional in the end as it is quite difficult to debug.